requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-telegram-bot
load_dotenv